refactor track-bundle dir
