+++ packages/core/src/track-bundle/index.ts
@@ -1,1 +1,2 @@
 export * from './track-bundle.types';
+export * from './track-bundle.utils';
